make usage more friendly; add README
We now use command-line parameters, there's an option to write to an
output file instead of purely using a temporary directory, and the
temporary directory is properly deleted at the end. (Errors are ignored
in the delete operation, so if OS file locking is such that the file
can't be deleted while the PDF viewer is looking at it, there will be no
complaints.)

Simple README.md added with summary of the program, usage details, and
dependencies.